Create demo and update constrain_rate
The constrain_rate function was using linear indices into the Q matrix,
that is, the indices would only work if Q was flattened (or raveled)
first. This commit changes that so instead the indices are tuples
of arrays as would be returned by np.nonzero. (They could also be tuples
of tuples/lists or potentially lists of tuples/lists, though this was
not the intended format.)